# Improve filter performance for long object lists for 'in' literals