--- a/CHANGES.rst
+++ b/CHANGES.rst
@@ -7,6 +7,7 @@ Version 3.16.1 (2020-10-??)
 
 * Fix an issue where :func:`~kartothek.io.dask.dataframe.collect_dataset_metadata` would execute
   ``get_parquet_metadata`` at graph construction time
+* Improve performance for "in" predicate literals using long object lists as values
 
 Version 3.16.0 (2020-09-29)
 ===========================