# Description:

For long lists of strings in the `in` clause this should help speed up the filtering significantly. I've seen this dominating query planing and IO. While this is not necessarily an encouraged practice, I've seen users doing this rather often.

With this change we can see several orders of magnitude better performance for very large lists (>>100). In particular, the new implementation essentially scales with the size of the array but is approximately constant in the length of the value. This is probably only true as long as `len(value) << len(array)` but this should be true for most applications.

For native types the performance of native numpy is not beatable and therefore I build in the switch based on types. While the new implementation seems to be slightly slower for small samples even for string types I consider this overhead to be a price worth paying to save on complexity. Hitting this threshold properly would otherwise require a lot of measurements. The optimal cutoff is probably a function of both array and value length. This might also be dependent on the kind of strings of the input dataset since the comparison of strings is also not constant time. Therefore I'm fine for this being slower.

W.r.t. the below measurements, I was meassuring them by enable/disabling the new code path explicitly. Therefore the code I was benchmarking used an `if enabled` instead of the `if value.dtype.kind in ("U", "S", "O")` but this should not impact the benchmark.

cc @NeroCorleone you have been looking into this as well

| | | array_size / enabled | | | | | |
| --------- | ------------- | ------------- | --------------- | --------------- | ---------------- | ---------------- | ----------------- |
|   column  |  filter_size  |  10000 / True |  10000 / False  |  100000 / True  |  100000 / False  |  1000000 / True  |  1000000 / False
|   int64   |       10      |  3.82±0.06ms  |     216±10μs    |    15.8±0.3ms   |     660±20μs     |     246±10ms     |     7.26±0.1ms |
|   int64   |      100      |   3.65±0.2ms  |     907±30μs    |    15.8±0.6ms   |   10.4±0.04ms    |     237±7ms      |      154±1ms |
|   int64   |      1000     |   3.90±0.2ms  |     978±60μs    |    16.3±0.4ms   |    11.5±0.2ms    |     244±6ms      |      168±1ms |
|  unicode  |       10      |   5.48±0.2ms  |    2.32±0.1ms   |    51.9±0.9ms   |     41.6±1ms     |     980±10ms     |     1.10±0.02s |
|  unicode  |      100      |   5.38±0.2ms  |     21.2±1ms    |     52.5±1ms    |     475±30ms     |     987±30ms     |     10.2±0.2s |
|  unicode  |      1000     |   5.97±0.3ms  |     208±6ms     |     54.6±2ms    |    4.05±0.07s    |     973±9ms      |       timeout |