@lr4d 2020-10-09
Interesting benchmarks, @fjetter . Thanks for the nice description.
Did you look into potentially using categoricals for this?