@fjetter 2020-10-09
> Did you look into potentially using categoricals for this?

No, not yet. Performance should only be marginally better for cases where the size of the to-be-filtered array is very large. The only way how we'd safe time/memory here is that the join would probably be a bit cheaper. otherwise I'd expect categoricals to be slower since they require one additional pass over the data and require more complicated internals. Overall I figured the complexity wouldn't be worth it